requirements.txt
xlwings
formulas
openpyxl
numpy
psycopg2
python-dotenv
google-generativeai